requirements.txt
requests
PyJWT
boto3
botocore
cryptography
PyJWT
requests
tweepy
certifi
chardet
urllib3
six
idna
cffi
oauthlib
pycparser
gspread
oauth2client